+++ src/main/kotlin/com/pestphp/pest/PestTestRunLineMarkerProvider.kt
@@ -1,6 +1,7 @@
 package com.pestphp.pest
 
 import com.intellij.execution.lineMarker.RunLineMarkerContributor
+import com.intellij.icons.AllIcons.RunConfigurations.TestState.Run
 import com.intellij.psi.PsiElement
 import com.jetbrains.php.lang.lexer.PhpTokenTypes
 import com.jetbrains.php.lang.psi.PhpPsiUtil
@@ -8,14 +9,16 @@ import com.jetbrains.php.lang.psi.elements.impl.FunctionReferenceImpl
 
 class PestTestRunLineMarkerProvider : RunLineMarkerContributor() {
     override fun getInfo(leaf: PsiElement): Info? {
-        if (!PhpPsiUtil.isOfType(leaf, PhpTokenTypes.IDENTIFIER)) {
-            return null
-        }
+        if (!PhpPsiUtil.isOfType(leaf, PhpTokenTypes.IDENTIFIER)) return null
 
-        if (leaf.parent !is FunctionReferenceImpl || !leaf.parent.isPestTestReference()) {
-            return null
-        }
+        val reference = leaf.parent
+        if (reference !is FunctionReferenceImpl || !reference.isPestTestReference()) return null
 
-        return withExecutorActions(PestIcons.RUN_SINGLE_TEST)
+        val fqn = reference.toPestFqn()
+
+        val icon = fqn.firstOrNull { getTestStateIcon(it, leaf.project, false) !== Run }
+            ?.let { getTestStateIcon(it, leaf.project, false) } ?: Run
+
+        return withExecutorActions(icon)
     }
 }